Library Reference

Interfaces

TraitInterfaces.Interfaces.Algorithms.rename — Method

This is boilerplate for the systematic renaming of types/operations in an interface.

TraitInterfaces.Interfaces.Algorithms.sortcheck — Method

Throw an error if a the head of an AlgTerm (which refers to a term constructor) has arguments of the wrong sort. Returns the sort of the term.

TraitInterfaces.Interfaces.Syntax — Module

The building blocks that make up an interface

TraitInterfaces.Interfaces.Syntax.AlgAccessor — Type

AlgAccessor

The arguments to a type/term constructor serve a dual function as both arguments and also methods to extract the value of those arguments.

E.g. declaring Hom(dom::Ob, codom::Ob)::TYPE implicitly creates projection operations like dom(h::Hom)::Ob.

TraitInterfaces.Interfaces.Syntax.AlgAxiom — Type

Declare two (context dependent) terms are equal

TraitInterfaces.Interfaces.Syntax.AlgFunction — Type

A term constructor which is purely derivative on other term constructors, such as "square(x) := x * x" (where * is some other term constructor). One need not specify the implementation of such an operation when declaring an implementation.

TraitInterfaces.Interfaces.Syntax.AlgSort — Type

AlgSort

A sort, which is essentially a type constructor without arguments

TraitInterfaces.Interfaces.Syntax.AlgSort — Method

A type constructor, alternatively thought of as a (possibly-dependent) type where we have forgotten the arguments it depends on. E.g. given a type Matrix{3,4} for 3 × 4 matrices, the corresponding sort is Matrix.

TraitInterfaces.Interfaces.Syntax.AlgTerm — Type

Subtyped by TermVar or TermApp

TraitInterfaces.Interfaces.Syntax.Judgment — Type

Judgments constitute the content of an interface

TraitInterfaces.Interfaces.Syntax.TermApp — Type

A term, which is a variable name (Symbol) in some context or the application of an operation to some other terms.

TraitInterfaces.Interfaces.Syntax.TermConstructor — Type

Declare an operation

TraitInterfaces.Interfaces.Syntax.TermVar — Type

Reference something in the context by name

TraitInterfaces.Interfaces.Syntax.TrmTypConstructor — Type

Either a TypeConstructor or a TermConstructor

TraitInterfaces.Interfaces.Syntax.TypeApp — Type

A (possibly dependent) type, represented as a type constructor applied to some terms. The type can depend on some concrete data (terms) as well as some types.

TraitInterfaces.Interfaces.Syntax.TypeConstructor — Type

Declare a type (possibly dependent on other types and terms)

TraitInterfaces.Interfaces.Syntax.TypeScope — Type

A context that a term (or type) can live in. It has an (ordered) list of names bound to types. Later types can refer to earlier types in the list.

The lookup field caches a quick way of getting the type from its name. There are no name collisions allowed.

TraitInterfaces.Interfaces.Syntax.VarArgType — Type

A type that is only allowed to be the last argument in a signature th

TraitInterfaces.Interfaces.InterfaceData — Module

Defining the data structure which stores data of an interface

TraitInterfaces.Interfaces.InterfaceData.Interface — Type

An interface has types, operations, and equations.

These may be abstract. An implementation provides concrete (Julia) types and methods for any abstract types and operations in the theory.

We can ascribe aliases to various types or operations.

Operations can be overloaded, meaning we can distinguish operations that have the same name if they have different type signatures. This should be used with caution, as implementations of the interface which assign the same concrete type to distinct abstract types can then cause method ambiguities.

TraitInterfaces.Interfaces.InterfaceData.Interface — Method

Create an empty interface - just provide the name of the interface

Base.union! — Method

Combine judgments from two interfaces, removing duplicates

TraitInterfaces.Interfaces.InterfaceData.allnames — Method

All type/term constructors as well as accessors

TraitInterfaces.Interfaces.InterfaceData.lookup — Function

From a name, get a type or an operation.

sorts not required unless there is overloading

TraitInterfaces.Interfaces.InterfaceModules.Dispatch — Type

The Dispatch implementation of an interface defers to type-dispatch.

TraitInterfaces.Interfaces.InterfaceModules.InitialModel′ — Type

The Initial model assigns Union{} to all AlgSorts. There is one implementation for any given theory.

TraitInterfaces.Interfaces.InterfaceModules.TerminalModel′ — Type

The Terminal model assigns Nothing to all AlgSorts. There is one implementation for any given theory.

TraitInterfaces.Interfaces.InterfaceModules.Trait — Type

Trait is a wrapper around a Julia value in order to signify it is being treated as a trait, i.e. a Holy trait, to be used as the first parameter in order to control dispatch.

TraitInterfaces.Interfaces.InterfaceModules.wrapper — Method

Given an interface module, take a name (and an abstract type, optionally) and construct a type with that name which plays the role of wrapping a model of the interface.

TraitInterfaces.Interfaces.InterfaceModules.@interface — Macro

Declare an interface (creates a module), possibly by extending an old one

TraitInterfaces.Interfaces.InterfaceModules.@theory — Macro

Same thing as @interface

Implementations

TraitInterfaces.Implementations.Check — Module

Runtime checking of method table to see if a trait has implemented a particular interface

TraitInterfaces.Implementations.Check.impl_types — Method

If m implements a GAT with a type constructor (identified by ident id), mapped to a Julia type, this function returns that Julia type.

TraitInterfaces.Interfaces.InterfaceModules.implements — Function

Check whether a model implements a particular theory.

If no types are provided, then we look up whether or not impl_type methods exist for this model + theory. If not, we will get a MethodError and assume that the model does not implement the theory. (WARNING: occasionally one has a complex type, such as foo(Int,String) which itself leads to a MethodError, and this can be confusing because it looks like the model doesn't implement the theory at all rather than just being an error in how it was implemented).

Once types are provided, we can check whether the theory is implemented by checking for each term constructor whether or not the model implements that (handled by a different implements method).

TraitInterfaces.Interfaces.InterfaceModules.implements — Method

Machine-friendly access to checking if a model implements a particular operation. The types vector is in bijection with the AlgSorts of the whole theory.

TraitInterfaces.Interfaces.InterfaceModules.implements — Method

User-friendly access to checking if a model implements an operation.

Throws an error if the name is overloaded. Anything programmatic should be calling a method which accepts method Idents rather than Symbols.

TraitInterfaces.Implementations.ModelInterface — Module

Any Julia value can be a trait which implements an interface, I. A trait t::T is considered to be implementing I iff, for all operations in I f(::A,::B,...), we have hasmethod(f, (Trait{M}, A, B)) == true.

TraitInterfaces.Implementations.ModelInterface.MissingMethodImplementation — Type

An error to throw when an implementation of an interface fails to implement a particular method

TraitInterfaces.Implementations.ModelInterface.get_judgment_runtime — Method

We need to look at the method after having forgotten the dependent types and still be able to tell what method was being implemented. Ambiguity can arise. Check there is only one possibility given the sorts provided.

TraitInterfaces.Implementations.ModelInterface.impl_type_declaration — Method

Register the concrete Julia type associated with a given interface, type in that interface, and model which implements that type.

E.g. if the interface ThCategory has a type Ob then

@instance ThCategory{Ob=Foo, Hom=Bar} [model::Baz] should generate the following code:

impl_type(::Baz, ThCategory.Ob) = Foo

TraitInterfaces.Implementations.ModelInterface.parse_instance_body — Method

Parses a raw julia block expression in @instance into JuliaFunctions

TraitInterfaces.Implementations.ModelInterface.qualify_function — Method

Does some preprocessing on the user-written method:

Add Trait param first (it shouldn't have it already).

qualify method name to be in theory module.

If an interface ThCategory with operation id(x::Ob)::Hom is implemented via:

@instance ThCategory{Ob=Foo, Hom=Bar} [model::Baz] begin 
  id(x::Foo)::Bar = ...
end

... then we syntactically modify this method to be:

ThCategory.id(model::Trait{<:Bar}, x::Foo) = ...

TraitInterfaces.Implementations.ModelInterface.typecheck_runtime — Method

Check that an instance properly implements some interface method.

E.g. if the interface ThCategory has a method compose(f::Hom, g::Hom)

then @instance ThCategory{Ob=Foo, Hom=Bar} [model::Baz] should generate the following code, to throw an error if the body of the macro didn't provide a method of the expected type:

hasmethod(ThCategory.compose, (Trait{<:Baz}, Bar, Bar)) || throw(
  MissingMethodImplementation(...))

TraitInterfaces.Implementations.ModelInterface.@instance — Macro

Usage: (TODO)

TraitInterfaces.Implementations.ModelInterface.@withmodel — Macro

Automatically add Trait trait parameter to some specified methods in a code block.

TraitInterfaces.Implementations.SpecialModels — Module

Any interface can be implemented in some canonical ways:

Just using dispatch

Assigning all types in the interface the empty (Union{}) type

Assigning all types in the interface the singleton (Nothing) type

TraitInterfaces.Implementations.SpecialModels.InitialModel — Constant

The unique term of type InitialModel′

TraitInterfaces.Implementations.SpecialModels.TerminalModel — Constant

The unique term of type TerminalModel′