Library Reference

Interfaces

TraitInterfaces.Interfaces.Algorithms.rename — Method

This is boilerplate for the systematic renaming of types/operations in an interface.

TraitInterfaces.Interfaces.Algorithms.sortcheck — Method

Throw an error if a the head of an AlgTerm (which refers to a term constructor) has arguments of the wrong sort. Returns the sort of the term.

TraitInterfaces.Interfaces.Syntax — Module

The building blocks that make up an interface

TraitInterfaces.Interfaces.Syntax.AlgAccessor — Type

AlgAccessor

The arguments to a type/term constructor serve a dual function as both arguments and also methods to extract the value of those arguments.

E.g. declaring Hom(dom::Ob, codom::Ob)::TYPE implicitly creates projection operations like dom(h::Hom)::Ob.

TraitInterfaces.Interfaces.Syntax.AlgAxiom — Type

Declare two (context dependent) terms are equal

TraitInterfaces.Interfaces.Syntax.AlgFunction — Type

A term constructor which is purely derivative on other term constructors, such as "square(x) := x * x" (where * is some other term constructor). One need not specify the implementation of such an operation when declaring an implementation.

TraitInterfaces.Interfaces.Syntax.AlgSort — Type

AlgSort

A sort, which is essentially a type constructor without arguments

TraitInterfaces.Interfaces.Syntax.AlgSort — Method

A type constructor, alternatively thought of as a (possibly-dependent) type where we have forgotten the arguments it depends on. E.g. given a type Matrix{3,4} for 3 × 4 matrices, the corresponding sort is Matrix.

TraitInterfaces.Interfaces.Syntax.AlgTerm — Type

Subtyped by TermVar or TermApp

TraitInterfaces.Interfaces.Syntax.Judgment — Type

Judgments constitute the content of an interface

TraitInterfaces.Interfaces.Syntax.TermApp — Type

A term, which is a variable name (Symbol) in some context or the application of an operation to some other terms.

TraitInterfaces.Interfaces.Syntax.TermConstructor — Type

Declare an operation

TraitInterfaces.Interfaces.Syntax.TermVar — Type

Reference something in the context by name

TraitInterfaces.Interfaces.Syntax.TrmTypConstructor — Type

Either a TypeConstructor or a TermConstructor

TraitInterfaces.Interfaces.Syntax.TypeApp — Type

A (possibly dependent) type, represented as a type constructor applied to some terms. The type can depend on some concrete data (terms) as well as some types.

TraitInterfaces.Interfaces.Syntax.TypeConstructor — Type

Declare a type (possibly dependent on other types and terms)

TraitInterfaces.Interfaces.Syntax.TypeScope — Type

A context that a term (or type) can live in. It has an (ordered) list of names bound to types. Later types can refer to earlier types in the list.

The lookup field caches a quick way of getting the type from its name. There are no name collisions allowed.

TraitInterfaces.Interfaces.Syntax.VarArgType — Type

A type that is only allowed to be the last argument in a signature th

TraitInterfaces.Interfaces.InterfaceData — Module

Defining the data structure which stores data of an interface

TraitInterfaces.Interfaces.InterfaceData.Interface — Type

An interface has types, operations, and equations.

These may be abstract. An implementation provides concrete (Julia) types and methods for any abstract types and operations in the theory.

We can ascribe aliases to various types or operations.

Operations can be overloaded, meaning we can distinguish operations that have the same name if they have different type signatures. This should be used with caution, as implementations of the interface which assign the same concrete type to distinct abstract types can then cause method ambiguities.

TraitInterfaces.Interfaces.InterfaceData.Interface — Method

Create an empty interface - just provide the name of the interface

Base.union! — Method

Combine judgments from two interfaces, removing duplicates

TraitInterfaces.Interfaces.InterfaceData.allnames — Method

All type/term constructors as well as accessors

TraitInterfaces.Interfaces.InterfaceData.lookup — Function

From a name, get a type or an operation.

sorts not required unless there is overloading

TraitInterfaces.Interfaces.InterfaceModules.Dispatch — Type

The Dispatch implementation of an interface defers to type-dispatch.

TraitInterfaces.Interfaces.InterfaceModules.Trait — Type

Trait is a wrapper around a Julia value in order to signify it is being treated as a trait, i.e. a Holy trait, to be used as the first parameter in order to control dispatch.

TraitInterfaces.Interfaces.InterfaceModules.wrapper — Method

Given an interface module, take a name (and an abstract type, optionally) and construct a type with that name which plays the role of wrapping a model of the interface.

TraitInterfaces.Interfaces.InterfaceModules.@interface — Macro

Declare an interface (creates a module), possibly by extending an old one

TraitInterfaces.Interfaces.InterfaceModules.@theory — Macro

Same thing as @interface

Implementations

Modules = [
  TraitInterfaces.Implementations.Check,
  TraitInterfaces.Implementations.ModelInterface,
  TraitInterfaces.Implementations.SpecialModels,
]